Test automation homework fixes
